# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Talent Management System",
        "description": "TMS Employee Resource Skills Tracking Management System"
    },
    "paths": {
...
        "/upadateemployeetracking": {
            "put": {
                "tags": [
                    "TMS Employee Details"
                ],
                "operationId": "upadateemployeetracking",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "updateEmployeeSkills",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/upadateemployeetracking"
                          }
                    }
                   
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/updateEmployeeDataByHR": {
            "put": {
                "tags": [
                    "TMS Employee Details"
                ],
                "operationId": "updateEmployeeDataByHR",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "updateEmployeeDataByHR",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateEmployee"
                          }
                    }
                   
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "updateEmployee": {
            "type": "object",
            "properties": {
                "integraExperience": {
                    "format":"float",
                    "type": "number"
                },
                "ModifiedDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "RoleId": {
                    "format": "int32",
                    "type": "integer"
                },
                "PrimarySkill": {
                    "format": "int32",
                    "type": "integer"
                },
                "IsReportingHead": {
                    "format": "int32",
                    "type": "integer"
                },
                "ModifiedBy": {
                    "type": "string"
                   
                },
                "Availability": {
                    "format": "int32",
                    "type": "integer"
                },
                "EmployeeId": {
                    "type": "string"
                }
            }
        },
...
        "upadateemployeetracking": {
            "type": "object",
            "properties": {
               "Id": {
                    "format": "int32",
                    "type": "integer"
                },
                "UserId": {
                    "format": "int32",
                    "type": "integer"
                },
                "ToDate": {
                    "type": "string"
                }
            }
        },
...
    }
}